refactor: コード調整
