+++ GitHubRepositorySearch/GitHubRepositorySearch/DataLayer/Store/RepoStore/RepoStore.swift
@@ -70,6 +70,26 @@ extension RepoStore {
         try await addValues(response.items)
         return response
     }
+        
+    func fetchUserRepos(
+        userName: String,
+        accessToken: String?,
+        sort: String?,
+        direction: String?,
+        perPage: Int?,
+        page: Int?
+    ) async throws -> ListResponse<Repo> {
+        let response = try await gitHubAPIClient.fetchUserRepos(
+            userName: userName,
+            accessToken: accessToken,
+            sort: sort,
+            direction: direction,
+            perPage: perPage,
+            page: page
+        )
+        try await addValues(response.items)
+        return response
+    }
     
     func fetchStarredRepos(
         userName: String,
@@ -87,7 +107,6 @@ extension RepoStore {
             perPage: perPage,
             page: page
         )
-//        try await addValues(response.starredRepos.map { StarredRepoTranslator.translate(from: $0) })
         try await addValues(response.starredRepos.map { $0.repo })
         return response
     }
