+++ GitHubRepositorySearch/GitHubRepositorySearch/Presentation/Profile/LoginUserView/LoginUserViewState.swift
