+++ GitHubRepositorySearch/GitHubRepositorySearch/Presentation/Profile/LoginView/LoginViewState.swift
