+++ GitHubRepositorySearch/GitHubRepositorySearch/Presentation/Search/RepoDetailsView/RepoDetailsViewState.swift
@@ -1,0 +1,154 @@
+//
+//  RepoDetailsViewState.swift
+//  GitHubRepositorySearch
+//
+//  Created by HIROKI IKEUCHI on 2025/05/09.
+//
+
+import Foundation
+import IKEHGitHubAPIClient
+
+@MainActor
+@Observable
+final class RepoDetailsViewState {
+    
+    // MARK: - Property
+    
+    let repoID: Repo.ID
+    let tokenStore: TokenStore
+    let loginUserStore: LoginUserStore
+    let repoStore: RepoStore
+    let starredRepoStore: StarredRepoStore
+
+    var isFetchingStarred: Bool = false
+    private var isUpdatingStarred: Bool = false
+    
+    var error: Error?
+    
+    var repo: Repo? {
+        repoStore.valuesDic[repoID]
+    }
+    
+    var loginUser: LoginUser? {
+        loginUserStore.loginUser
+    }
+    
+    var disableStarButton: Bool {
+        (loginUser == nil) ||
+        isFetchingStarred ||
+        isUpdatingStarred
+    }
+    
+    var isStarred: Bool {
+        starredRepoStore.valuesDic[repoID]?.isStarred ?? false
+    }
+    
+    // MARK: - LifeCycle
+        
+    init(
+        repoID: Repo.ID,
+        tokenStore: TokenStore = TokenStore.shared,
+        loginUserStore: LoginUserStore = LoginUserStore.shared,
+        repoStore: RepoStore = RepoStore.shared,
+        starredRepoStore: StarredRepoStore = StarredRepoStore.shared,
+    ) {
+        self.repoID = repoID
+        self.tokenStore = tokenStore
+        self.loginUserStore = loginUserStore
+        self.repoStore = repoStore
+        self.starredRepoStore = starredRepoStore
+    }
+}
+
+// MARK: - Actions
+
+extension RepoDetailsViewState {
+    func onAppear() {
+        Task {
+            await checkIfRepoIsStarred()
+        }
+    }
+    
+    func handleStarButtonTapped() {
+        // すでに処理中の場合は何もしない
+        if isUpdatingStarred {
+            return
+        }
+        Task {
+            await updateStarred()
+        }
+    }
+}
+
+// MARK: - Star Methods
+
+extension RepoDetailsViewState {
+    
+    private func checkIfRepoIsStarred() async {
+        // 必要な情報の確認
+        guard let repo,
+              loginUser != nil,
+              let accessToken = await tokenStore.accessToken
+        else {
+            return
+        }
+        // 状態の更新
+        isFetchingStarred = true
+        defer {
+            isFetchingStarred = false
+        }
+        
+        // スター状態の取得
+        do {
+            try await starredRepoStore.checkIsRepoStarred(
+                repoID: repo.id,
+                accessToken: accessToken,
+                owner: repo.owner.login,
+                repo: repo.name,
+            )
+        } catch {
+            self.error = error
+        }
+    }
+        
+    private func updateStarred() async {
+        
+        // 必要な情報のチェック
+        guard let repo,
+              let accessToken = await tokenStore.accessToken
+        else {
+            return
+        }
+        
+        // 状態の更新
+        isUpdatingStarred = true
+        defer {
+            isUpdatingStarred = false
+        }
+        // スター状態を更新
+        let currentIsStarred = isStarred
+        let currentStarsCount = repo.starsCount
+        do {
+            if isStarred {
+                // スターを取り消す
+                // 一時的に値を更新する
+                try await repoStore.updateStarsCountInLocal(repoID: repo.id, starsCount: max(currentStarsCount - 1, .zero)) // スター数は0未満にならない
+                try await starredRepoStore.updateStarredInLocal(repoID: repoID, isStarred: false)
+                // 実際の更新処理
+                try await starredRepoStore.unstarRepo(repoID: repoID, accessToken: accessToken, owner: repo.owner.login, repo: repo.name)
+            } else {
+                // スターをつける
+                // 一時的に値を更新する
+                try await repoStore.updateStarsCountInLocal(repoID: repo.id, starsCount: currentStarsCount + 1)
+                try await starredRepoStore.updateStarredInLocal(repoID: repoID, isStarred: true)
+                // 実際の更新処理
+                try await starredRepoStore.starRepo(repoID: repoID, accessToken: accessToken, owner: repo.owner.login, repo: repo.name)
+            }
+        } catch {
+            // スターの状態をもとに戻す
+            try? await starredRepoStore.updateStarredInLocal(repoID: repoID, isStarred: currentIsStarred)
+            try? await repoStore.updateStarsCountInLocal(repoID: repo.id, starsCount: currentStarsCount)
+            self.error = error
+        }
+    }
+}
